pyproject.toml
[project]
name = "study-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.86.0",
    "openai-agents>=0.0.17",
    "python-dotenv",
]
